handle NIC transmission and ARP request
To verify the transmission of the arp request works as expected, we capture
the packets in/out of the NIC by the '-object filter-dump,...' qemu command line option and then load the
packets dump into wireshark to inspect the captured in/out packets. From the dump, we see the arp packets sent and
arp replies (alghouth the software does not process the arp reply yet).