make usb boot work in qemu
We encountered a weird issue regarding usb. When we access a USB drive from user space via syscall, we get stuck.
The root cause is we call msleep when waiting for a event TRB from the controller. The msleep will never return since
the interrupt is disabled when entering kernel from user space (because of InterruptGate). Thus either we stuck forever
in the hlt instruction or in the loop waiting for someone increasing the tick (which will not happen).

One tricky part is, if we add printf here or there, we may get luckly that the stuck went away. The reason is,
printf may trigger vga memory movement, which takes time. Thansk to printf, the event TRB may be ready before we
enter the msleep loop (in ConsumerTRBRing::dequeue) waiting for the readiness. Adding a busy loop has the same effect.