one line change: set the DIR bit of StatusStageTRB correctly
This is a bug that bothers me for a few weeks. Is causes stuck when I run SOS on my toshiba laptop.
The symptom is, after I put TRBs for the GetDescriptor request to the transfer ring and ring the doorbell,
I don't get an event back in the event ring nor do I get an interrupt (even though I have enabled interrupts - the code is on a local branch. I may commit them later).

Since the hardware does not give any hint of what's going wrong and the code works on QEMU.., I have to guess what's the root cause.

At first, I thought maybe it's caused by CPU cache. i.e., when I read/write DMA memory or memory mapped IO region, cache
hides the access from the memory/device registers. But even if I tweak the paging and disable cache for those memory regions,
the issue is still there.

Finally, I carefully read Haiku OS's xhci driver and found the way I setup DIR bit of the StatusStageTRB is wrong.
It's a bit tricky since when there is a data transfer, the DIR of StatusStageTRB actually is opposite to the DIR of DataStageTRB.
XHCI spec table '4-7' precisely decibe the rules of setting the DIR bits for StatusStageTRB. But unfortunately I didn't pay
enough attention to this part earlier... The doc is too long ... 600+ pages.