Fix a weird issue that dump paging cause issue in XHCI
I encountered a very weired bug that if I dump kernel pgdir in setup_paging, I encounter usb error
later. The issue only happens on the real hardware but not Qemu.

It turns out that after I start the XHCI controller in XHCIDriver::reset, the port with the device connected
is not in the connected state immediately. So the port reset logic is skipped thus the port is left in an 'uninitialized'
state. I tried to inject some delay and it works. But I'm not sure how much delay is absolutely necessary. My current fix is
just wait until I found some port's CCS (current connect status) bit is set.

The trigger of the issue (dump kernel pgdir) and the effect of the issue (some USB error) seems completely unreleated.
My best guess is, dump kernel pgdir warms ups TLB. This makes the XHCI controller and port reset be faster. But the hardware
need some time to be fully ready. Previously when we don't dump the kernel pgdir, TLB miss causes some delays so it happens to
work well. But after all, it's just my guess..