rename mymain back to main
For main function, the compiler by default apply '-mpreferred-stack-boundary=4'. The %esp will be aligned to 16 bytes boundary and the old esp will be stored in register like %ebx. The syscall assembly function used to clobber %eax/%ebx/%ecx/%edx/%esi/%edi without saving them first. This cause a wrong %ebx being used to access the stack after returning from a syscall.

The workaround of renaming main to mymain works since the compiler seems not enforce '-mpreferred-stack-boundary=4' for a function called mymain. The stack is still being accessed thru %esp/%ebp which always have proper value.

Another workaround is to force '-mpreferred-stack-boundary=2' when compiling the file containing the main function. Note the valid range of '-mpreferred-stack-boundary' is between 2 and 12. This way, the compiled assembly does not need another register like %ebx to store the old stack. We use %esp/%ebp directly to access the stack.

But the ideal solution which is also the one I implemented is to save all the clobbered registers before doing a syscall and restore then afterwards.